# Some optimizations for sending UDP messages