--- a/gelf/writer_test.go
+++ b/gelf/writer_test.go
@@ -5,9 +5,12 @@
 package gelf
 
 import (
+	"compress/flate"
 	"crypto/rand"
 	"encoding/base64"
 	"fmt"
+	"io"
+	"io/ioutil"
 	"strings"
 	"testing"
 	"time"
@@ -226,5 +229,22 @@ func TestExtraData(t *testing.T) {
 			t.Errorf("_line didn't roundtrip (%v != %v)", int(msg.Extra["_line"].(float64)), extra["_line"].(int))
 			return
 		}
+	}
+}
+
+func BenchmarkWriteBestSpeed(b *testing.B) {
+	r, err := NewReader("127.0.0.1:0")
+	if err != nil {
+		b.Fatalf("NewReader: %s", err)
+	}
+	go io.Copy(ioutil.Discard, r)
+	w, err := NewWriter(r.Addr())
+	if err != nil {
+		b.Fatalf("NewWriter: %s", err)
+	}
+	w.CompressionType = flate.BestSpeed
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		w.Write([]byte("This is a message"))
 	}
 }