--- a/gelf/writer_test.go
+++ b/gelf/writer_test.go
@@ -265,3 +265,20 @@ func BenchmarkWriteNoCompression(b *testing.B) {
 		w.Write([]byte("This is a message"))
 	}
 }
+
+func BenchmarkWriteDisableCompressionCompletely(b *testing.B) {
+	r, err := NewReader("127.0.0.1:0")
+	if err != nil {
+		b.Fatalf("NewReader: %s", err)
+	}
+	go io.Copy(ioutil.Discard, r)
+	w, err := NewWriter(r.Addr())
+	if err != nil {
+		b.Fatalf("NewWriter: %s", err)
+	}
+	w.CompressionType = CompressNone
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		w.Write([]byte("This is a message"))
+	}
+}